pyproject.toml
[project]
name = "portfolio"
version = "1.1.0"
description = "Dependency tree for Iliyan's portfolio web app"

dependencies = [
    "Django>=5.2.4",
    "mockito>=1.5.4",
    "setuptools>=80.9.0",
    "email-validator>=2.2.0",
    "pyyaml>=6.0.2",
    "django-bootstrap5>=25.1",
    "pillow>=11.3.0",
    "django-libsass>=0.9",
    "django-compressor>=4.5.1",
    "markdown>=3.8.2",
    "uvicorn>=0.35.0",
    "uvicorn-worker>=0.3.0",
    "gunicorn>=23.0.0",
    "typing-extensions>=4.14.1",
    "whitenoise>=6.9.0",
    "pip>=25.2",
    "django-csp>=4.0",
    "loguru>=0.7.3",
    "loki-logger-handler>=1.1.2",
    "requests>=2.32.5",
    "psycopg2-binary>=2.9.10",
    "django-parler>=2.3",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "GPL"}

[dependency-groups]
dev = [
    "python-dotenv>=1.1.1",
]
test = [
    "black>=25.1.0",
    "django-migration-linter>=5.2.0",
    "django-test-migrations>=1.5.0",
    "fb-sapp>=0.5.8",
    "flake8>=7.3.0",
    "pylint>=3.3.8",
    "pyre-check>=0.9.25",
    "setuptools-rust>=1.11.1",
]
